tests/wpt/web-platform-tests/css/css-align/gaps/gap-animation-003.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Box Alignment Test: Default gap test animation</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://www.w3.org/TR/css-align-3/#column-row-gap">
<link rel="help" href="https://www.w3.org/TR/css-animations-1/#keyframes">
<meta name="assert" content="This test checks that the default value for gap property, which is 'normal', is not interpolable.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @keyframes gap-to-100 {
    to { gap: 100px; }
  }

  #target {
    animation-name: gap-to-100;
    animation-duration: 2s;
    animation-delay: -1s;
    animation-play-state: paused;
  }
</style>
<body>
  <div id="target"></div>
  <div id="log"></div>

  <script>
    test(
      function(){
        var target = document.getElementById("target");
        assert_equals(getComputedStyle(target).gap, "100px");
        assert_equals(getComputedStyle(target).rowGap, "100px");
        assert_equals(getComputedStyle(target).columnGap, "100px");
      }, "Default gap is not interpolable");
  </script>
</body>
