Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-align/gaps/gap-animation-003-expected.txt


FAIL Default gap is not interpolable assert_equals: expected "100px" but got "100px 100px"
